+++ crates/anodized-fuzz/src/lib.rs
@@ -1,1 +1,1 @@
-pub use anodized_macros::fuzz_fn;
+pub use anodized_macros::fuzz_fn_call;
